[#3917] Fix determining the oldest SST file for intest cleanup
Summary:
RocksDB allows deleting only the oldest SST files.
During intents cleanup, we try to determine the oldest SST file using its largest hybrid time.
But sometimes we could multiple files with the same largest hybrid time (when one of them contains only one operation).
So we could incorrectly determine the oldest SST file and attempt to delete it will result in the following log spew:
```
W0310 02:23:41.136687  6102 db_impl.cc:5756] T 17bf282e30194315bb93ddba1128c5fc P b955edf0aa3242e88690f6c744a8e096 [I]: DeleteFile /009235.sst failed --- target file in level 0 must be the oldest.
I0310 02:23:41.136925  6102 tablet.cc:721] T 17bf282e30194315bb93ddba1128c5fc P b955edf0aa3242e88690f6c744a8e096: Intents SST file will be deleted: { total_size: 66525 base_size: 66341 uncompressed_size: 66698 name: "/009235.sst" db_path: "/mnt/d0/yb-data/tserver/data/rocksdb/table-5fc67ff38b1e4e9dbc8a86b2aac8ad0d/tablet-17bf282e30194315bb93ddba1128c5fc.intents" imported: 0 being_compacted: 0 column_family_name: default level: 0 smallest: { seqno: 1125901475261132 user_frontier: 0x00000000912b22a0 -> { op_id: 579.174348811 hybrid_time: { physical: 1583788957617741 } history_cutoff: <invalid> hybrid_time_filter: <invalid> } } largest: { seqno: 1125901475261140 user_frontier: 0x0000000033ae7ce0 -> { op_id: 579.174348811 hybrid_time: { physical: 1583788957617741 } history_cutoff: <invalid> hybrid_time_filter: <invalid> } } }, max ht: { physical: 1583788957617741 }, min running transaction start ht: <max>
```

This diff fixes the issue by switching to seqno instead of hybrid time for detemining oldest SST file.
Added handling of deletion failure. So we would not end up in spamming a log in case of failure.
Also added ability to view RocksDB state from TServer web interface.

Test Plan:
ybd --gtest_filter QLTransactionTest.DeleteFlushedIntents
{F13406}

Reviewers: mikhail, bogdan

Reviewed By: bogdan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8121

+++ src/yb/tablet/tablet.cc
@@ -546,7 +546,7 @@ Result<bool> Tablet::IntentsDbFlushFilter(const rocksdb::MemTable& memtable) {
       std::chrono::steady_clock::now() > memtable.FlushStartTime() + timeout) {
     rocksdb::FlushOptions options;
     options.wait = false;
-    regular_db_->Flush(options);
+    RETURN_NOT_OK(regular_db_->Flush(options));
   }
 
   return false;
@@ -716,12 +716,13 @@ void Tablet::DoCleanupIntentFiles() {
     const rocksdb::LiveFileMetaData* best_file = nullptr;
     files.clear();
     intents_db_->GetLiveFilesMetaData(&files);
+    auto min_largest_seq_no = std::numeric_limits<rocksdb::SequenceNumber>::max();
     for (const auto& file : files) {
-      auto& frontier = down_cast<docdb::ConsensusFrontier&>(*file.largest.user_frontier);
-      auto file_max_ht = frontier.hybrid_time();
-      if (file_max_ht < best_file_max_ht) {
+      if (file.largest.seqno < min_largest_seq_no) {
+        min_largest_seq_no = file.largest.seqno;
+        auto& frontier = down_cast<docdb::ConsensusFrontier&>(*file.largest.user_frontier);
+        best_file_max_ht = frontier.hybrid_time();
         best_file = &file;
-        best_file_max_ht = file_max_ht;
       }
     }
 
@@ -734,8 +735,17 @@ void Tablet::DoCleanupIntentFiles() {
         << "Intents SST file will be deleted: " << best_file->ToString()
         << ", max ht: " << best_file_max_ht << ", min running transaction start ht: "
         << min_running_start_ht;
-    regular_db_->Flush(rocksdb::FlushOptions());
-    intents_db_->DeleteFile(best_file->name);
+    auto flush_status = regular_db_->Flush(rocksdb::FlushOptions());
+    if (!flush_status.ok()) {
+      LOG_WITH_PREFIX(WARNING) << "Failed to flush regular db: " << flush_status;
+      break;
+    }
+    auto delete_status = intents_db_->DeleteFile(best_file->name);
+    if (!delete_status.ok()) {
+      LOG_WITH_PREFIX(WARNING) << "Failed to delete " << best_file->ToString()
+                               << ", all files " << AsString(files) << ": " << delete_status;
+      break;
+    }
   }
 
   if (best_file_max_ht != HybridTime::kMax) {
@@ -1621,7 +1631,7 @@ Status Tablet::Flush(FlushMode mode, FlushFlags flags, int64_t ignore_if_flushed
   }
 
   if (flush_intents && mode == FlushMode::kSync) {
-    intents_db_->WaitForFlush();
+    RETURN_NOT_OK(intents_db_->WaitForFlush());
   }
 
   return Status::OK();
@@ -2242,7 +2252,7 @@ void Tablet::FlushIntentsDbIfNecessary(const yb::OpId& lastest_log_entry_op_id) 
             << FLAGS_num_raft_ops_to_force_idle_intents_db_to_flush << " is allowed";
         rocksdb::FlushOptions options;
         options.wait = false;
-        intents_db_->Flush(options);
+        WARN_NOT_OK(intents_db_->Flush(options), "Flush intents db failed");
       }
     }
   }
@@ -2537,7 +2547,7 @@ void Tablet::ForceRocksDBCompactInTest() {
     docdb::ForceRocksDBCompact(regular_db_.get());
   }
   if (intents_db_) {
-    intents_db_->Flush(rocksdb::FlushOptions());
+    CHECK_OK(intents_db_->Flush(rocksdb::FlushOptions()));
     docdb::ForceRocksDBCompact(intents_db_.get());
   }
 }
